♻️ move locale view model
