+++ lib/ui/boulder_area/widgets/boulder_area_details_list_tab.dart
@@ -1,9 +1,9 @@
 import 'package:breizh_blok_mobile/data/data_sources/api/model/request_strategy.dart';
 import 'package:breizh_blok_mobile/domain/models/boulder_area/boulder_area.dart';
-import 'package:breizh_blok_mobile/ui/core/view_models/boulder_bloc.dart';
-import 'package:breizh_blok_mobile/ui/core/view_models/boulder_filter_bloc.dart';
-import 'package:breizh_blok_mobile/ui/core/view_models/boulder_filter_grade_bloc.dart';
-import 'package:breizh_blok_mobile/ui/core/view_models/boulder_order_bloc.dart';
+import 'package:breizh_blok_mobile/ui/boulder/view_models/boulder_bloc.dart';
+import 'package:breizh_blok_mobile/ui/boulder/view_models/boulder_filter_bloc.dart';
+import 'package:breizh_blok_mobile/ui/boulder/view_models/boulder_filter_grade_bloc.dart';
+import 'package:breizh_blok_mobile/ui/boulder/view_models/boulder_order_bloc.dart';
 import 'package:breizh_blok_mobile/ui/core/widgets/boulder_list_builder.dart';
 import 'package:breizh_blok_mobile/ui/core/widgets/downloads_area_button.dart';
 import 'package:flutter/material.dart';
