+++ lib/ui/core/widgets/boulder_list_builder_sort_button.dart
@@ -1,7 +1,7 @@
 import 'package:breizh_blok_mobile/data/data_sources/api/model/api_order_choice.dart';
 import 'package:breizh_blok_mobile/data/data_sources/api/model/api_order_param.dart';
 import 'package:breizh_blok_mobile/i18n/app_localizations.dart';
-import 'package:breizh_blok_mobile/ui/core/view_models/boulder_order_bloc.dart';
+import 'package:breizh_blok_mobile/ui/boulder/view_models/boulder_order_bloc.dart';
 import 'package:breizh_blok_mobile/ui/core/widgets/sort_button.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
