+++ lib/ui/profile/widgets/profile_screen.dart
@@ -1,6 +1,6 @@
 import 'package:breizh_blok_mobile/i18n/app_localizations.dart';
 import 'package:breizh_blok_mobile/ui/download/widgets/downloaded_boulder_areas_screen.dart';
-import 'package:breizh_blok_mobile/ui/profile/widgets/profile_language_switcher.dart';
+import 'package:breizh_blok_mobile/ui/locale/widgets/locale_switch.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
@@ -27,7 +27,7 @@ class ProfileScreen extends StatelessWidget {
               context.pushNamed(DownloadedBoulderAreasScreen.route.name);
             },
           ),
-          ProfileLanguageSwitcher(
+          LocaleSwitch(
             currentLocale: Localizations.localeOf(context).languageCode,
           ),
         ],
