+++ test/ui/boulder/view_models/boulder_bloc_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.5 from annotations
-// in breizh_blok_mobile/test/ui/core/view_models/boulder_bloc_test.dart.
+// in breizh_blok_mobile/test/ui/boulder/view_models/boulder_bloc_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
