+++ test/ui/locale/view_models/locale_bloc_view_model_test.dart
@@ -1,31 +1,31 @@
 // ignore_for_file: avoid_print
 
 import 'package:breizh_blok_mobile/i18n/app_localizations.dart';
-import 'package:breizh_blok_mobile/ui/core/view_models/locale_bloc.dart';
+import 'package:breizh_blok_mobile/ui/locale/view_models/locale_view_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 void main() {
-  group('LocaleBloc', () {
-    Future<LocaleBloc> pumpLocaleBlocPlayground(WidgetTester tester) async {
+  group('LocaleViewModel', () {
+    Future<LocaleViewModel> pumpPlayground(WidgetTester tester) async {
       WidgetsFlutterBinding.ensureInitialized();
-      final localeBloc = await LocaleBloc.create();
+      final localeBloc = await LocaleViewModel.create();
       await tester.pumpWidget(
-        BlocProvider<LocaleBloc>(
+        BlocProvider<LocaleViewModel>(
           create: (context) => localeBloc,
           child: Builder(
             builder: (context) {
               return MaterialApp(
                 localizationsDelegates: AppLocalizations.localizationsDelegates,
                 supportedLocales: AppLocalizations.supportedLocales,
-                locale: context.watch<LocaleBloc>().state.locale,
+                locale: context.watch<LocaleViewModel>().state.locale,
                 home: Column(
                   children: [
                     Text(
                       // ignore: lines_longer_than_80_chars
-                      'locale: ${context.watch<LocaleBloc>().state.locale.languageCode}',
+                      'locale: ${context.watch<LocaleViewModel>().state.locale.languageCode}',
                     ),
                   ],
                 ),
@@ -47,7 +47,7 @@ void main() {
         tester.platformDispatcher.localeTestValue = const Locale('fr');
         SharedPreferences.setMockInitialValues({});
 
-        final localeBloc = await pumpLocaleBlocPlayground(tester);
+        final localeBloc = await pumpPlayground(tester);
 
         expect(find.text('locale: fr'), findsOneWidget);
         expect(localeBloc.state.locale, const Locale('fr'));
@@ -63,7 +63,7 @@ void main() {
         SharedPreferences.setMockInitialValues({});
         tester.platformDispatcher.localeTestValue = const Locale('en');
 
-        final localeBloc = await pumpLocaleBlocPlayground(tester);
+        final localeBloc = await pumpPlayground(tester);
 
         expect(find.text('locale: en'), findsOneWidget);
         expect(localeBloc.state.locale, const Locale('en'));
@@ -79,7 +79,7 @@ void main() {
         tester.platformDispatcher.localeTestValue = const Locale('FOO');
         SharedPreferences.setMockInitialValues({});
 
-        final localeBloc = await pumpLocaleBlocPlayground(tester);
+        final localeBloc = await pumpPlayground(tester);
 
         expect(find.text('locale: fr'), findsOneWidget);
         expect(localeBloc.state.locale, const Locale('fr'));
@@ -95,7 +95,7 @@ void main() {
         tester.platformDispatcher.localeTestValue = const Locale('fr');
         SharedPreferences.setMockInitialValues({kLocalePrefs: 'en'});
 
-        final localeBloc = await pumpLocaleBlocPlayground(tester);
+        final localeBloc = await pumpPlayground(tester);
 
         final prefs = await SharedPreferences.getInstance();
         expect(prefs.getString(kLocalePrefs), 'en');
@@ -112,7 +112,7 @@ void main() {
         tester.platformDispatcher.localeTestValue = const Locale('fr');
         SharedPreferences.setMockInitialValues({kLocalePrefs: 'foo'});
 
-        final localeBloc = await pumpLocaleBlocPlayground(tester);
+        final localeBloc = await pumpPlayground(tester);
 
         final prefs = await SharedPreferences.getInstance();
         expect(prefs.getString(kLocalePrefs), 'foo');
